improvements to first cargo choice

+++ com-mmxlabs-lng-scheduler/com.mmxlabs.scheduler.optimiser/src/com/mmxlabs/scheduler/optimiser/schedule/timewindowscheduling/LadenRouteData.java
@@ -5,13 +5,13 @@
 package com.mmxlabs.scheduler.optimiser.schedule.timewindowscheduling;
 
 public class LadenRouteData {
-	public final long ladenTimeAtMaxSpeed;
-	public final long ladenTimeAtNBOSpeed;
+	public final int ladenTimeAtMaxSpeed;
+	public final int ladenTimeAtNBOSpeed;
 	public final long ladenRouteCost;
 	public final long ladenRouteDistance;
 	public final int transitTime;
 	
-	public LadenRouteData(final long ladenTimeAtMaxSpeed, final long ladenTimeAtNBOSpeed, final long ladenRouteCost, final long ladenRouteDistance, final int transitTime) {
+	public LadenRouteData(final int ladenTimeAtMaxSpeed, final int ladenTimeAtNBOSpeed, final long ladenRouteCost, final long ladenRouteDistance, final int transitTime) {
 		this.ladenTimeAtMaxSpeed = ladenTimeAtMaxSpeed;
 		this.ladenTimeAtNBOSpeed = ladenTimeAtNBOSpeed;
 		this.ladenRouteCost = ladenRouteCost;
@@ -23,8 +23,8 @@ public class LadenRouteData {
 		return Long.compare(a.ladenTimeAtMaxSpeed, b.ladenTimeAtMaxSpeed);
 	}
 	
-	public static long getMinimumTravelTime(LadenRouteData[] data) {
-		long minimumTravelTime = Long.MAX_VALUE;
+	public static int getMinimumTravelTime(LadenRouteData[] data) {
+		int minimumTravelTime = Integer.MAX_VALUE;
 		for (LadenRouteData ladenRouteData : data) {
 			if (ladenRouteData.ladenTimeAtMaxSpeed < minimumTravelTime) {
 				minimumTravelTime = ladenRouteData.ladenTimeAtMaxSpeed;
